+++ frontend/components/Requests/NewRequest.js
@@ -82,6 +82,7 @@ export default function NewRequest() {
       alert("Start Date cannot be after or the same as End Date");
     } else {
       try {
+
         //post request
         instance
           .post(`http://localhost:8080/api/service-requests/create`, {
@@ -115,6 +116,8 @@ export default function NewRequest() {
 
   return (
     <React.Fragment>
+
+    
       <Paper
         sx={{
           p: 2,
