+++ frontend/pages/Customer/NewRequest/index.js
@@ -1,15 +1,14 @@
 // import statements
 import * as React from "react";
+import NewRequest from "../../../components/Requests/NewRequest";
+import CustomerDash from "../index";
 
-import NewRequest from "../../../components/NewRequest/NewRequest";
-import CustomerDash from "./index";
-import Account from "../../../components/Account/Account";
 
 
 export default function NewRequestScreen() {
   return (
     <CustomerDash>
-      <Account />
+      <NewRequest />
     </CustomerDash>
   );
 }
