+++ frontend/pages/Customer/account.js
@@ -3,6 +3,7 @@ import * as React from "react";
 import Account from "../../components/Account/Account";
 import CustomerDash from "./index";
 
+
 export default function AccountScreen() {
   return (
     <CustomerDash>
